Add config for testing api disconnections
If the API disconnects for some reason, your node may be unable to
get a remote value. That remote value will then 'stick' and the
node may get stuck in an undesired state.

This example shows that when the API connection is unavailable, the
LED on the board turns on. This could be modified to switch between
remote and local temperature sensors, or perform fallover logic.

Note that using esphome to remotely log the node will count as an
API connection in addition to the connection to Home Assistant.